from ("%u") to ("%d",(int)) for size_t
svn path=/trunk/iCub/; revision=15995

+++ src/libraries/skinDynLib/src/dynContact.cpp
@@ -233,7 +233,7 @@ void dynContact::setVerbose(unsigned int verb){
 bool dynContact::checkVectorDim(const Vector &v, unsigned int dim, const string &descr){
     if(v.length() != dim){
         if(verbose)
-            fprintf(stderr, "Error in dynContact: unexpected dimension of vector %s, %u\n", descr.c_str(), v.length());
+            fprintf(stderr, "Error in dynContact: unexpected dimension of vector %s, %d\n", descr.c_str(), (int)v.length());
         return false;
     }
     return true;
